+++ sistema.estudiante/src/module-info.java
@@ -8,4 +8,5 @@
 module sistema.estudiante {
 	requires java.desktop;
 	requires java.logging;
+	requires java.sql;
 }
